+++ app/build.gradle
@@ -33,8 +33,7 @@ android {
 dependencies {
 
     implementation project(path: ':cieidsdk')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.61"
-    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'androidx.multidex:multidex:2.0.1'
