+++ app/src/main/res/layout/activity_redirection.xml
