+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
     <string name="app_name">CieIDSDKSampleApp</string>
     <string name="waiting">WAITING CIE...</string>
+    <string name="internal">Flusso integrato</string>
+    <string name="redirection">Flusso con reindirizzamento</string>
 </resources>
