+++ cieidsdk/build.gradle
@@ -15,9 +15,9 @@ android {
 
         //ambienti
         //PRODUZIONE
-        //buildConfigField "String", "BASE_URL_IDP", "\"https://idserver.servizicie.interno.gov.it/idp/\""
+        buildConfigField "String", "BASE_URL_IDP", "\"https://idserver.servizicie.interno.gov.it/idp/\""
         //COLLAUDO
-        buildConfigField "String", "BASE_URL_IDP", "\"https://preproduzione.idserver.servizicie.interno.gov.it/idp/\""
+        //buildConfigField "String", "BASE_URL_IDP", "\"https://preproduzione.idserver.servizicie.interno.gov.it/idp/\""
 
     }
 
@@ -33,14 +33,14 @@ android {
 
 
 ext {
-    okhttp_version = '4.0.1'
-    retrofit_version = '2.6.0'
-    kotlin_version = '1.3.61'
+    okhttp_version = '4.9.0'
+    retrofit_version = '2.9.0'
+    kotlin_version = '1.4.10'
 }
 
 dependencies {
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 
     //network
     implementation "com.squareup.okhttp3:okhttp:$okhttp_version"
@@ -48,13 +48,12 @@ dependencies {
     implementation "com.squareup.retrofit2:converter-scalars:$retrofit_version"
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
     implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofit_version"
-    implementation 'com.squareup.okio:okio:2.4.3'
+    implementation 'com.squareup.okio:okio:2.8.0'
 
     //rxjava
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.18'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
 
-    implementation 'io.kotlintest:kotlintest:2.0.7'
 
 }
